// src/models/Colector.ts
/**
 *
 * @export
 * @interface Colector
 */
export interface Colector {
  /**
   *
   * @type {number}
   * @memberof Colector
   */
  id: number;
  /**
   *
   * @type {string}
   * @memberof Colector
   */
  nombre: string;
  /**
   *
   * @type {string}
   * @memberof Colector
   */
  apellido: string;
  /**
   *
   * @type {string}
   * @memberof Colector
   */
  identificacion: string;
  /**
   *
   * @type {number}
   * @memberof Colector
   */
  telefono: number;
}